third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-prefixed-after-unprefixed.html

<!DOCTYPE html>
<html>
    <head>
        <title>Prefixed and unprefixed EME API</title>
        <script src="encrypted-media-utils.js"></script>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>
        <video id="testVideo"></video>
        <div id="log"></div>
        <p>Call prefixed EME API after unprefixed EME API.</p>
        <script>
            async_test(function(test)
            {
                var video = document.getElementById('testVideo');
                assert_not_equals(video, null);

                function loadStarted()
                {
                    MediaKeys.create('org.w3.clearkey').then(function(mediaKeys) {
                        assert_not_equals(mediaKeys, null);
                        assert_equals(mediaKeys.keySystem, 'org.w3.clearkey');
                        return video.setMediaKeys(mediaKeys);
                    }).then(function(result) {
                        assert_throws('InvalidStateError',
                                      function () { video.webkitGenerateKeyRequest('webkit-org.w3.clearkey'); });
                        test.done();
                    }).catch(function(error) {
                        forceTestFailureFromPromise(test, error);
                    });
                }

                video.src = '../content/test-encrypted.webm';
                waitForEventAndRunStep('loadstart', video, loadStarted, test);
            }, 'Call prefixed EME API after unprefixed EME API.');
        </script>
    </body>
</html>
